+++ app/src/main/java/com/dsource/idc/jellowintl/activities/SpeechEngineBaseActivity.java
@@ -39,6 +39,7 @@ import static com.dsource.idc.jellowintl.utility.SessionManager.KHA_IN;
 import static com.dsource.idc.jellowintl.utility.SessionManager.KN_IN;
 import static com.dsource.idc.jellowintl.utility.SessionManager.ML_IN;
 import static com.dsource.idc.jellowintl.utility.SessionManager.MR_IN;
+import static com.dsource.idc.jellowintl.utility.SessionManager.MR_TTS;
 import static com.dsource.idc.jellowintl.utility.SessionManager.PA_IN;
 import static com.dsource.idc.jellowintl.utility.SessionManager.TA_IN;
 import static com.dsource.idc.jellowintl.utility.SessionManager.TE_IN;
@@ -117,6 +118,7 @@ public class SpeechEngineBaseActivity extends BaseActivity{
             case BN_IN:
             case BE_IN:
             case MR_IN:
+            case MR_TTS:
             case ES_ES:
             case TA_IN:
             case DE_DE:
@@ -144,6 +146,7 @@ public class SpeechEngineBaseActivity extends BaseActivity{
             case BN_IN:
             case BE_IN:
             case MR_IN:
+            case MR_TTS:
             case ES_ES:
             case TA_IN:
             case DE_DE:
@@ -172,6 +175,7 @@ public class SpeechEngineBaseActivity extends BaseActivity{
             case BN_IN:
             case BE_IN:
             case MR_IN:
+            case MR_TTS:
             case ES_ES:
             case TA_IN:
             case DE_DE:
@@ -405,11 +409,11 @@ public class SpeechEngineBaseActivity extends BaseActivity{
         return SessionManager.NoTTSLang.contains(getSession().getLanguage());
     }
 
-    public void initiateSpeechEngineWithLanguage(String voice){
+    public void initiateSpeechEngineWithLanguage(String voice, String language){
         if(voice == null || voice.isEmpty()){
            voice = getAvailableVoicesForLanguage(getSession().getLanguage()).split(",")[0];
         }
-        setupSpeechEngine(voice, getSession().getLanguage());
+        setupSpeechEngine(voice, language);
     }
 
     static HashMap<String, String> voiceGender = new HashMap<String, String>(){
@@ -461,8 +465,7 @@ public class SpeechEngineBaseActivity extends BaseActivity{
             put("fr-fr-x-frd-local"," (M)");
             put("fr-fr-x-vlf-local"," (F)");
             put("gu-in-x-guf-local"," (F)");
-            put("gu-in-x-gum-local"," (F)");
-            put("hi-in-x-cfn-local"," (F)");
+            put("gu-in-x-gum-local"," (M)");
             put("hi-in-x-hia-local"," (F)");
             put("hi-in-x-hic-local"," (F)");
             put("hi-in-x-hid-local"," (M)");